SCOTT: Shrink the Atari slow-draw op's delay field
a8_byte_to_write carried the OPCODE_DELAY frame count as an int. The
value is at most arg0(0..31) * 20 = 620, so a uint16_t holds it with
room to spare. With the int gone the struct packs from 12 to 8 bytes —
a third off every recorded op in the slow-draw queue.

Behaviour unchanged; saga vector test 6/6 (apple2 + atari), ASan-clean.

+++ terps/scott/saga/atari_8bit_vector_draw.c
@@ -94,7 +94,9 @@ typedef struct {
     uint8_t value90;
     uint8_t valueA0;
     bool fill_bg;
-    int delay;
+    // Frame count for OPCODE_DELAY ops; at most arg0(0..31) * 20 = 620, so a
+    // uint16_t holds it with room to spare and keeps the struct compact.
+    uint16_t delay;
 } a8_byte_to_write;
 
 static a8_byte_to_write *bytes_to_write = NULL;
